moved tokenizer and (refactored) vectorizer
